Muevo código de las clases derivadas de Query a la base
    refs:#90

    refs:#90
