+++ series_tiempo_ar_api/apps/api/query/query.py
@@ -62,7 +62,7 @@ class Query(object):
             self.add_collapse(collapse=periodicity)
 
         periodicity = get_periodicity_human_format(field_model.distribution.periodicity)
-        self.es_query.add_series(name, rep_mode, periodicity=periodicity)
+        self.es_query.add_series(name, rep_mode, periodicity)
 
     @staticmethod
     def get_max_periodicity(periodicities):
