+++ series_tiempo_ar_api/apps/api/tests/es_query_tests.py
@@ -22,7 +22,8 @@ class QueryTest(TestCase):
     start_date = '2010-01-01'
     end_date = '2015-01-01'
 
-    series_month = 'random_series-month-0'
+    single_series = 'random_series-month-0'
+    rep_mode = 'value'
     series_periodicity = 'month'  # Periodicidad de la serie anterior
 
     @classmethod
@@ -37,20 +38,20 @@ class QueryTest(TestCase):
         self.assertFalse(self.query.series)
 
     def test_pagination(self):
-        self.query.add_series(self.series_month, periodicity=self.series_periodicity)
+        self.query.add_series(self.single_series, self.rep_mode, self.series_periodicity)
         self.query.add_pagination(self.start, self.limit)
         data = self.query.run()
 
         self.assertEqual(len(data), self.limit - self.start)
 
     def test_pagination_limit(self):
-        self.query.add_series(self.series_month, periodicity=self.series_periodicity)
+        self.query.add_series(self.single_series, self.rep_mode, self.series_periodicity)
         self.query.add_pagination(self.start, self.max_limit)
         data = self.query.run()
         self.assertEqual(len(data), self.max_limit - self.start)
 
     def test_time_filter(self):
-        self.query.add_series(self.series_month, periodicity=self.series_periodicity)
+        self.query.add_series(self.single_series, self.rep_mode, self.series_periodicity)
         self.query.add_filter(self.start_date, self.end_date)
         data = self.query.run()
         for row in data:
@@ -66,19 +67,19 @@ class QueryTest(TestCase):
             self.assertLessEqual(date, end_date)
 
     def test_execute_single_series(self):
-        self.query.add_series(self.series_month, periodicity=self.series_periodicity)
+        self.query.add_series(self.single_series, self.rep_mode, self.series_periodicity)
         data = self.query.run()
 
         self.assertTrue(data)
 
     def test_default_return_limits(self):
-        self.query.add_series(self.series_month, periodicity=self.series_periodicity)
+        self.query.add_series(self.single_series, self.rep_mode, self.series_periodicity)
         data = self.query.run()
 
         self.assertEqual(len(data), self.default_limit)
 
     def test_add_series(self):
-        self.query.add_series(self.series_month, periodicity=self.series_periodicity)
+        self.query.add_series(self.single_series, self.rep_mode, self.series_periodicity)
         data = self.query.run()
 
         self.assertTrue(data)
@@ -86,8 +87,8 @@ class QueryTest(TestCase):
         self.assertTrue(len(data[0]) == 2)
 
     def test_add_two_series(self):
-        self.query.add_series(self.series_month, 'value', periodicity=self.series_periodicity)
-        self.query.add_series(self.series_month,
+        self.query.add_series(self.single_series, 'value', periodicity=self.series_periodicity)
+        self.query.add_series(self.single_series,
                               'percent_change',
                               periodicity=self.series_periodicity)
         data = self.query.run()
@@ -96,6 +97,10 @@ class QueryTest(TestCase):
         # Expected: rows de 3 datos: timestamp, serie 1, serie 2
         self.assertTrue(len(data[0]) == 3)
 
+    @raises(QueryError)
+    def test_try_collapse(self):
+        self.query.add_collapse(interval='year', agg='avg')
+
 
 class CollapseQueryTests(TestCase):
 
@@ -103,6 +108,7 @@ class CollapseQueryTests(TestCase):
     limit = 15
 
     single_series = 'random_series-month-0'
+    rep_mode = 'value'
     series_periodicity = 'month'
 
     @classmethod
@@ -120,13 +126,13 @@ class CollapseQueryTests(TestCase):
         self.assertFalse(data)
 
     def test_execute_single(self):
-        self.query.add_series(self.single_series)
+        self.query.add_series(self.single_series, self.rep_mode, self.series_periodicity)
 
         data = self.query.run()
         self.assertTrue(data)
 
     def test_start_limit(self):
-        self.query.add_series(self.single_series)
+        self.query.add_series(self.single_series, self.rep_mode, self.series_periodicity)
         self.query.add_pagination(self.start, self.limit)
         data = self.query.run()
 
@@ -134,7 +140,7 @@ class CollapseQueryTests(TestCase):
 
     def test_init_from_other(self):
         other_query = ESQuery(index=settings.TEST_INDEX)
-        other_query.add_series(self.single_series, periodicity=self.series_periodicity)
+        other_query.add_series(self.single_series, self.rep_mode, self.series_periodicity)
         self.query = CollapseQuery(settings.TEST_INDEX, other=other_query)
         self.query.add_collapse()
         data = self.query.run()
@@ -144,7 +150,7 @@ class CollapseQueryTests(TestCase):
         """Testea que luego de agregar un collapse default, los
         resultados sean anuales, es decir cada uno a un año de
         diferencia con su anterior"""
-        self.query.add_series(self.single_series)
+        self.query.add_series(self.single_series, self.rep_mode, self.series_periodicity)
         self.query.add_collapse()
         data = self.query.run()
         prev_timestamp = None
@@ -159,7 +165,7 @@ class CollapseQueryTests(TestCase):
             prev_timestamp = parsed_timestamp
 
     def test_collapse_custom_params(self):
-        self.query.add_series(self.single_series)
+        self.query.add_series(self.single_series, self.rep_mode, self.series_periodicity)
         self.query.add_collapse(interval='quarter')
         data = self.query.run()
         prev_timestamp = None
@@ -175,7 +181,7 @@ class CollapseQueryTests(TestCase):
 
     def test_add_two_collapses(self):
         """Esperado: El segundo collapse overridea el primero"""
-        self.query.add_series(self.single_series)
+        self.query.add_series(self.single_series, self.rep_mode, self.series_periodicity)
         self.query.add_collapse(interval='quarter')
         self.query.add_collapse(interval='year')
         data = self.query.run()
@@ -193,7 +199,7 @@ class CollapseQueryTests(TestCase):
 
     def test_init_from_other_collapse_query(self):
         other_query = CollapseQuery(index=settings.TEST_INDEX)
-        other_query.add_series(self.single_series)
+        other_query.add_series(self.single_series, self.rep_mode, self.series_periodicity)
         self.query = CollapseQuery(index=settings.TEST_INDEX, other=other_query)
         data = self.query.run()
         self.assertTrue(data)
